moved intervalEstimation
