+++ output/CLT_1.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group E,associate's degree,standard,completed,63,58,60
-female,group E,high school,free/reduced,none,50,59,53
-female,group C,high school,standard,completed,50,67,63
-female,group D,high school,standard,none,68,72,78
-male,group B,high school,standard,none,83,77,67
-male,group C,some college,standard,none,93,88,81
-male,group C,some college,standard,none,53,47,41
-male,group D,some high school,standard,none,78,84,81
-female,group C,high school,standard,none,63,83,75
-female,group C,high school,standard,none,57,66,68
-female,group B,associate's degree,standard,none,83,91,93
-female,group D,high school,standard,none,54,64,66
-male,group B,bachelor's degree,standard,none,67,66,60
-female,group C,some college,standard,none,53,74,59
-male,group C,some high school,standard,completed,64,63,63
-female,group E,high school,standard,none,74,75,74
-female,group D,associate's degree,standard,completed,59,62,64
-female,group C,some college,standard,none,68,75,77
-female,group D,some college,standard,none,66,70,70
-male,group D,high school,standard,completed,54,58,56
+female,group C,master's degree,standard,completed,96,100,100
+female,group E,some high school,free/reduced,none,58,57,53
+male,group D,associate's degree,standard,none,92,87,88
+male,group E,some high school,standard,completed,97,88,86
+male,group E,high school,standard,none,82,66,60
+male,group A,high school,standard,completed,72,70,65
+male,group B,associate's degree,standard,none,70,67,63
+male,group B,high school,standard,none,55,41,41
+male,group B,high school,standard,completed,85,85,85
+female,group D,high school,standard,none,64,74,74
+female,group E,master's degree,free/reduced,none,73,74,73
+male,group D,some college,free/reduced,none,80,71,76
+female,group C,some high school,free/reduced,none,44,56,53
+female,group D,some high school,standard,completed,81,90,87
+male,group C,bachelor's degree,standard,completed,67,65,60
+male,group A,bachelor's degree,standard,none,72,57,53
+male,group B,master's degree,standard,none,86,78,77
+female,group E,high school,standard,none,78,78,73
+female,group C,high school,standard,none,74,91,87
+female,group D,high school,standard,completed,73,96,96
