+++ output/CLT_3.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group A,some college,standard,none,71,86,80
-male,group B,high school,standard,none,72,72,62
-male,group C,bachelor's degree,free/reduced,none,77,87,85
-female,group D,some college,standard,none,65,63,70
-male,group B,high school,free/reduced,completed,46,47,49
-female,group B,some high school,free/reduced,none,53,64,63
-male,group A,associate's degree,standard,none,53,56,51
-male,group C,high school,standard,none,79,58,57
-male,group C,some college,standard,none,70,64,60
-female,group D,high school,standard,none,73,76,73
-male,group D,some college,standard,none,56,61,56
-female,group C,some college,standard,none,80,74,80
-male,group B,high school,standard,none,83,77,67
-female,group A,some high school,standard,none,68,76,71
-male,group D,some college,standard,none,88,93,93
-female,group C,high school,standard,completed,45,58,59
-female,group D,some college,standard,none,45,59,55
-male,group E,some high school,free/reduced,none,49,40,38
-female,group B,bachelor's degree,free/reduced,none,65,76,74
-male,group E,some college,free/reduced,completed,79,91,89
+female,group C,high school,standard,completed,59,64,64
+female,group D,some college,standard,none,60,66,63
+female,group C,associate's degree,standard,completed,67,83,88
+male,group B,master's degree,standard,none,81,67,67
+male,group A,some high school,standard,none,53,48,43
+male,group C,some college,standard,none,67,51,50
+male,group B,some college,standard,completed,52,47,55
+female,group C,associate's degree,free/reduced,completed,56,72,73
+female,group C,some college,standard,none,84,96,96
+male,group D,high school,standard,none,86,85,82
+female,group B,high school,standard,none,59,65,68
+female,group C,master's degree,standard,completed,70,76,76
+female,group C,some college,standard,completed,68,85,83
+male,group D,master's degree,standard,completed,96,82,89
+female,group C,master's degree,standard,none,78,78,88
+male,group C,bachelor's degree,standard,none,69,73,60
+male,group B,some college,standard,none,44,37,35
+male,group D,some high school,free/reduced,none,41,45,38
+male,group C,high school,standard,none,63,72,65
+female,group D,some high school,standard,none,81,89,90
