+++ output/CLT_4.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-male,group B,high school,standard,none,46,41,45
-male,group E,bachelor's degree,standard,none,77,61,59
-female,group E,associate's degree,free/reduced,none,48,52,58
-female,group D,some high school,standard,none,66,66,70
-male,group B,some college,free/reduced,completed,64,62,60
-male,group D,associate's degree,standard,completed,72,80,73
-female,group E,high school,standard,none,60,65,60
-female,group D,high school,standard,none,79,90,88
-female,group D,some college,standard,completed,52,57,63
-male,group C,some high school,standard,none,59,56,54
-male,group C,high school,standard,none,78,68,68
-male,group D,master's degree,free/reduced,completed,65,60,65
-male,group E,bachelor's degree,standard,none,76,75,72
-female,group A,some college,standard,none,71,86,80
-male,group D,high school,standard,none,74,74,70
-female,group D,high school,free/reduced,none,78,75,77
-male,group C,associate's degree,standard,none,66,68,64
-male,group E,bachelor's degree,standard,none,100,95,92
-female,group B,some high school,standard,none,64,69,69
-male,group D,some high school,standard,none,68,60,54
+female,group C,some college,free/reduced,completed,51,71,67
+female,group C,some college,standard,none,62,81,69
+male,group C,high school,free/reduced,none,57,59,55
+male,group B,some high school,standard,none,58,59,52
+female,group C,high school,standard,completed,50,67,63
+male,group A,some high school,standard,completed,85,82,81
+female,group D,associate's degree,free/reduced,completed,82,88,90
+female,group B,associate's degree,standard,none,51,55,55
+male,group E,master's degree,free/reduced,completed,92,80,85
+female,group D,associate's degree,standard,completed,74,85,88
+female,group D,master's degree,standard,none,78,91,88
+female,group B,some college,standard,none,61,66,65
+female,group B,some high school,standard,none,69,81,81
+female,group D,high school,free/reduced,completed,49,74,67
+female,group C,some high school,free/reduced,none,41,57,51
+male,group C,some college,standard,none,92,85,87
+male,group B,master's degree,free/reduced,none,69,70,70
+female,group B,some college,standard,none,82,87,86
+male,group B,associate's degree,standard,none,73,66,59
+female,group D,bachelor's degree,free/reduced,none,58,78,71
