+++ output/CLT_5.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-male,group E,bachelor's degree,standard,none,88,80,78
-female,group D,high school,free/reduced,none,69,92,91
-female,group E,high school,free/reduced,none,50,59,53
-male,group A,some college,standard,none,77,69,65
-male,group E,some college,free/reduced,none,67,74,70
-female,group E,bachelor's degree,free/reduced,none,83,75,74
-female,group B,some high school,standard,completed,65,66,72
-female,group C,some high school,standard,completed,62,73,78
-female,group E,bachelor's degree,standard,none,100,100,100
-male,group D,bachelor's degree,standard,none,74,68,78
-female,group E,high school,standard,none,79,75,75
-female,group D,associate's degree,standard,completed,74,85,88
-female,group C,high school,standard,none,53,58,57
-male,group E,associate's degree,standard,none,100,91,88
-male,group B,some high school,standard,none,58,56,52
-male,group E,high school,free/reduced,none,47,50,41
-female,group C,some college,standard,completed,69,78,76
-male,group C,associate's degree,standard,none,66,68,64
-female,group D,some high school,standard,none,78,78,83
-female,group D,some college,standard,none,77,77,78
+female,group B,associate's degree,standard,none,45,65,60
+female,group C,high school,standard,none,57,58,60
+female,group D,some high school,standard,completed,58,70,73
+female,group C,high school,standard,none,56,53,53
+female,group B,associate's degree,standard,completed,77,84,85
+male,group E,some high school,standard,completed,97,88,86
+male,group E,bachelor's degree,free/reduced,none,54,57,51
+female,group D,associate's degree,free/reduced,none,50,68,60
+male,group D,associate's degree,standard,none,83,73,68
+female,group B,bachelor's degree,standard,none,68,70,75
+female,group D,associate's degree,free/reduced,none,53,62,56
+male,group B,master's degree,standard,completed,90,96,98
+male,group C,some high school,free/reduced,completed,57,68,60
+female,group C,associate's degree,standard,completed,73,86,86
+male,group D,some high school,standard,completed,72,77,76
+male,group C,some college,standard,completed,72,81,80
+male,group B,some high school,free/reduced,completed,49,59,56
+female,group B,some college,standard,none,61,66,65
+male,group C,high school,standard,completed,81,77,77
+male,group D,some high school,standard,none,71,66,59
