+++ output/CLT_6.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-male,group C,bachelor's degree,free/reduced,none,68,55,52
-female,group D,some high school,standard,completed,53,64,67
-female,group D,bachelor's degree,standard,none,78,100,96
-male,group C,high school,free/reduced,completed,64,67,66
-female,group C,bachelor's degree,standard,none,92,86,94
-male,group E,associate's degree,free/reduced,completed,100,91,91
-male,group B,master's degree,standard,completed,100,100,98
-female,group C,bachelor's degree,standard,none,59,64,64
-male,group E,associate's degree,standard,none,73,62,66
-male,group D,some high school,standard,completed,69,63,60
-female,group E,bachelor's degree,standard,none,49,49,55
-male,group B,high school,standard,none,55,54,57
-male,group D,bachelor's degree,free/reduced,none,51,42,42
-female,group B,some college,standard,completed,42,64,59
-male,group E,associate's degree,free/reduced,none,59,49,49
-female,group B,some high school,standard,none,59,71,61
-male,group A,high school,standard,none,78,82,68
-male,group C,high school,standard,completed,81,72,69
-female,group D,some college,standard,none,78,79,89
-male,group C,some college,free/reduced,completed,63,54,58
+male,group B,high school,free/reduced,completed,46,47,49
+male,group C,bachelor's degree,standard,none,69,73,60
+female,group B,associate's degree,standard,none,64,67,70
+female,group A,high school,standard,completed,71,74,77
+female,group B,some college,standard,none,51,46,46
+female,group C,bachelor's degree,standard,none,57,56,60
+male,group B,high school,standard,completed,85,85,85
+male,group B,some college,standard,completed,69,68,69
+female,group D,master's degree,standard,none,78,91,88
+female,group C,master's degree,standard,none,78,78,88
+male,group C,high school,free/reduced,none,51,52,49
+male,group B,master's degree,standard,none,45,39,39
+female,group D,some high school,standard,completed,69,75,82
+male,group C,some college,standard,completed,64,62,63
+female,group B,high school,free/reduced,completed,59,65,63
+female,group C,bachelor's degree,standard,none,61,73,67
+female,group B,bachelor's degree,standard,none,42,49,50
+male,group D,some high school,standard,none,52,53,50
+female,group D,some college,free/reduced,none,77,80,82
+male,group B,high school,free/reduced,completed,53,60,57
