+++ output/CLT_7.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group D,some college,standard,none,65,63,70
-male,group C,high school,free/reduced,none,65,62,57
-male,group E,high school,standard,completed,93,82,82
-female,group C,some high school,free/reduced,none,41,57,51
-female,group C,some college,standard,completed,45,56,65
-male,group D,some high school,free/reduced,completed,49,54,56
-male,group B,some college,standard,none,52,53,48
-male,group C,bachelor's degree,standard,none,75,84,72
-female,group E,high school,standard,none,90,82,83
-male,group D,some college,free/reduced,none,82,69,73
-male,group D,high school,standard,none,83,71,68
-male,group D,some high school,standard,none,76,74,72
-male,group C,bachelor's degree,free/reduced,none,73,69,74
-male,group D,some college,standard,none,95,82,77
-male,group B,bachelor's degree,free/reduced,completed,60,68,60
-female,group C,some high school,free/reduced,none,70,77,79
-male,group D,bachelor's degree,free/reduced,completed,94,89,89
-female,group D,master's degree,free/reduced,none,67,75,82
-female,group E,some college,free/reduced,none,70,79,73
-male,group B,some high school,standard,none,78,62,59
+male,group C,some high school,standard,completed,68,56,64
+male,group C,high school,standard,none,61,61,59
+female,group C,high school,standard,none,56,53,53
+male,group C,associate's degree,free/reduced,none,76,67,63
+male,group D,some high school,standard,none,86,80,77
+female,group D,some college,free/reduced,completed,76,84,85
+female,group C,some college,standard,none,64,60,67
+female,group C,associate's degree,free/reduced,none,68,67,68
+female,group C,associate's degree,free/reduced,none,67,73,75
+female,group B,bachelor's degree,standard,completed,93,98,100
+female,group E,some high school,standard,none,59,63,62
+female,group A,associate's degree,standard,completed,76,94,83
+female,group B,master's degree,standard,none,81,97,89
+male,group C,some college,standard,none,63,64,61
+female,group D,some college,standard,completed,67,76,80
+male,group C,bachelor's degree,free/reduced,none,52,61,55
+male,group E,associate's degree,standard,none,95,77,74
+female,group C,high school,free/reduced,none,36,62,57
+female,group C,associate's degree,standard,none,38,44,44
+male,group B,some high school,standard,completed,75,72,75
