+++ output/CLT_9.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group C,high school,standard,completed,50,67,63
-female,group C,some college,standard,completed,70,89,88
-female,group C,some high school,standard,completed,69,75,81
-male,group A,high school,standard,none,59,54,50
-female,group E,some high school,standard,none,83,71,70
-male,group E,some college,free/reduced,completed,56,57,57
-female,group C,high school,standard,none,57,66,68
-female,group B,some high school,standard,none,53,56,59
-female,group D,some college,standard,none,85,97,94
-female,group D,associate's degree,free/reduced,completed,65,89,85
-female,group C,master's degree,standard,none,93,91,92
-female,group D,some high school,standard,completed,58,70,73
-female,group E,some college,free/reduced,completed,63,72,73
-male,group D,high school,standard,completed,100,89,95
-female,group B,high school,standard,none,39,47,46
-male,group B,some college,standard,completed,76,75,79
-male,group B,some high school,standard,none,58,56,52
-female,group C,associate's degree,free/reduced,none,68,67,68
-male,group D,high school,free/reduced,completed,53,59,54
-male,group D,bachelor's degree,standard,none,85,87,87
+female,group E,high school,standard,none,78,78,73
+male,group E,some high school,free/reduced,none,59,57,44
+male,group B,master's degree,standard,none,71,73,63
+male,group E,associate's degree,standard,none,88,76,67
+male,group E,bachelor's degree,free/reduced,none,49,48,46
+male,group D,high school,standard,none,58,54,49
+male,group E,bachelor's degree,standard,none,84,70,70
+male,group C,bachelor's degree,free/reduced,none,76,91,78
+male,group D,master's degree,standard,completed,82,90,95
+female,group D,bachelor's degree,free/reduced,none,71,89,86
+female,group E,bachelor's degree,standard,none,65,64,60
+male,group B,high school,standard,none,66,55,52
+female,group B,associate's degree,standard,none,62,67,68
+female,group B,some college,free/reduced,none,31,48,44
+female,group B,master's degree,standard,none,69,78,73
+female,group D,associate's degree,standard,completed,55,62,68
+male,group D,some high school,standard,completed,69,63,60
+male,group C,some high school,standard,none,43,42,39
+female,group A,some college,free/reduced,completed,78,97,95
+male,group C,some college,standard,completed,74,76,78
