+++ output/IE.csv
@@ -1,21 +1,21 @@
 gender,race/ethnicity,parental level of education,lunch,test preparation course,math score,reading score,writing score
-female,group C,some high school,free/reduced,none,41,57,51
-male,group D,some college,free/reduced,none,52,58,51
-female,group D,high school,free/reduced,completed,80,85,89
-female,group E,some high school,standard,none,94,97,95
-female,group B,associate's degree,free/reduced,none,44,56,51
-male,group B,high school,standard,completed,85,85,85
-male,group C,associate's degree,standard,completed,86,83,77
-male,group D,some high school,standard,completed,69,63,60
-female,group D,high school,standard,none,54,67,59
-male,group B,bachelor's degree,standard,none,84,77,79
-male,group C,associate's degree,standard,none,55,60,56
-female,group A,associate's degree,standard,completed,67,78,81
-male,group C,some college,standard,none,85,84,79
-male,group C,some high school,standard,completed,64,63,63
-male,group D,high school,standard,none,85,83,79
-male,group B,bachelor's degree,standard,none,65,62,59
-male,group C,some college,standard,completed,72,81,80
-female,group E,some college,standard,none,67,70,67
-female,group D,some high school,free/reduced,none,46,67,66
-female,group A,master's degree,standard,completed,74,91,92
+female,group C,high school,free/reduced,none,51,67,65
+female,group C,associate's degree,standard,completed,61,74,78
+male,group B,some high school,free/reduced,none,32,32,29
+male,group D,some high school,standard,none,73,78,82
+female,group C,associate's degree,free/reduced,none,44,62,53
+female,group C,some college,standard,none,64,60,67
+male,group C,some college,standard,none,92,85,87
+female,group D,bachelor's degree,standard,none,78,100,96
+male,group D,high school,standard,none,58,54,49
+female,group D,associate's degree,free/reduced,completed,74,87,87
+female,group B,bachelor's degree,free/reduced,none,54,69,68
+male,group E,associate's degree,standard,completed,85,81,80
+female,group B,associate's degree,free/reduced,none,54,67,63
+female,group C,some college,free/reduced,completed,51,71,67
+male,group D,associate's degree,standard,none,74,68,71
+female,group C,some high school,standard,none,61,76,68
+female,group D,associate's degree,free/reduced,completed,82,88,90
+female,group D,some college,standard,none,77,77,78
+male,group B,associate's degree,standard,completed,74,59,65
+male,group B,some high school,standard,none,68,75,68
